+++ dataforge-io/src/commonMain/kotlin/space/kscience/dataforge/io/IOFormat.kt
@@ -1,12 +1,13 @@
 package space.kscience.dataforge.io
 
-import io.ktor.utils.io.core.*
+import io.ktor.utils.io.core.Input
+import io.ktor.utils.io.core.Output
+import io.ktor.utils.io.core.readDouble
+import io.ktor.utils.io.core.writeDouble
 import space.kscience.dataforge.context.Context
 import space.kscience.dataforge.context.Factory
-import space.kscience.dataforge.io.IOFormat.Companion.NAME_KEY
 import space.kscience.dataforge.io.IOFormatFactory.Companion.IO_FORMAT_TYPE
 import space.kscience.dataforge.meta.Meta
-import space.kscience.dataforge.meta.MetaRepr
 import space.kscience.dataforge.misc.Named
 import space.kscience.dataforge.misc.Type
 import space.kscience.dataforge.names.Name
@@ -14,92 +15,53 @@ import space.kscience.dataforge.names.asName
 import kotlin.reflect.KType
 import kotlin.reflect.typeOf
 
+public fun interface IOReader<out T : Any> {
+
+    public fun readObject(input: Input): T
+}
+
+public fun interface IOWriter<in T : Any> {
+    public fun writeObject(output: Output, obj: T)
+}
+
+
 /**
  * And interface for reading and writing objects into with IO streams
  */
-public interface IOFormat<T : Any> : MetaRepr {
-    public val type: KType
+public interface IOFormat<T : Any> : IOReader<T>, IOWriter<T>
 
-    public fun writeObject(output: Output, obj: T)
-    public fun readObject(input: Input): T
+public fun <T : Any> Input.readObject(format: IOReader<T>): T = format.readObject(this@readObject)
 
-    public companion object {
-        public val NAME_KEY: Name = "name".asName()
-        public val META_KEY: Name = "meta".asName()
-    }
-}
-
-public fun <T : Any> Input.readWith(format: IOFormat<T>): T = format.readObject(this@readWith)
-
-public fun <T: Any> IOFormat<T>.readObject(binary: Binary): T = binary.read {
+public fun <T : Any> IOFormat<T>.readObjectFrom(binary: Binary): T = binary.read {
     readObject(this)
 }
 
 /**
  * Read given binary as object using given format
  */
-public fun <T : Any> Binary.readWith(format: IOFormat<T>): T = read {
-    readWith(format)
+public fun <T : Any> Binary.readWith(format: IOReader<T>): T = read {
+    readObject(format)
 }
 
-public fun <T : Any> Output.writeWith(format: IOFormat<T>, obj: T): Unit =
-    format.run { writeObject(this@writeWith, obj) }
+public fun <T : Any> Output.writeObject(format: IOWriter<T>, obj: T): Unit =
+    format.run { writeObject(this@writeObject, obj) }
 
-public inline fun <reified T : Any> IOFormat.Companion.listOf(
-    format: IOFormat<T>,
-): IOFormat<List<T>> = object : IOFormat<List<T>> {
-    override val type: KType = typeOf<List<T>>()
-
-    override fun writeObject(output: Output, obj: List<T>) {
-        output.writeInt(obj.size)
-        format.run {
-            obj.forEach {
-                writeObject(output, it)
-            }
-        }
-    }
-
-    override fun readObject(input: Input): List<T> {
-        val size = input.readInt()
-        return format.run {
-            List(size) { readObject(input) }
-        }
-    }
-
-    override fun toMeta(): Meta = Meta {
-        NAME_KEY put "list"
-        "contentFormat" put format.toMeta()
-    }
-
-}
-
-//public fun ObjectPool<Buffer>.fill(block: Buffer.() -> Unit): Buffer {
-//    val buffer = borrow()
-//    return try {
-//        buffer.apply(block)
-//    } catch (ex: Exception) {
-//        //recycle(buffer)
-//        throw ex
-//    }
-//}
 
 @Type(IO_FORMAT_TYPE)
-public interface IOFormatFactory<T : Any> : Factory<IOFormat<T>>, Named, MetaRepr {
+public interface IOFormatFactory<T : Any> : Factory<IOFormat<T>>, Named {
     /**
      * Explicit type for dynamic type checks
      */
     public val type: KType
 
-    override fun toMeta(): Meta = Meta {
-        NAME_KEY put name.toString()
-    }
-
     public companion object {
         public const val IO_FORMAT_TYPE: String = "io.format"
+        public val NAME_KEY: Name = "name".asName()
+        public val META_KEY: Name = "meta".asName()
     }
 }
 
-public fun <T : Any> IOFormat<T>.toBinary(obj: T): Binary = Binary { writeObject(this, obj) }
+public fun <T : Any> Binary(obj: T, format: IOWriter<T>): Binary = Binary { format.writeObject(this, obj) }
 
 public object DoubleIOFormat : IOFormat<Double>, IOFormatFactory<Double> {
     override fun build(context: Context, meta: Meta): IOFormat<Double> = this
@@ -114,20 +76,3 @@ public object DoubleIOFormat : IOFormat<Double>, IOFormatFactory<Double> {
 
     override fun readObject(input: Input): Double = input.readDouble()
 }
-
-//public object ValueIOFormat : IOFormat<Value>, IOFormatFactory<Value> {
-//    override fun invoke(meta: Meta, context: Context): IOFormat<Value> = this
-//
-//    override val name: Name = "value".asName()
-//
-//    override val type: KType get() = typeOf<Value>()
-//
-//    override fun writeObject(output: Output, obj: Value) {
-//        BinaryMetaFormat.run { output.writeValue(obj) }
-//    }
-//
-//    override fun readObject(input: Input): Value {
-//        return (BinaryMetaFormat.run { input.readMetaItem() } as? MetaItemValue)?.value
-//            ?: error("The item is not a value")
-//    }
-//}