+++ dataforge-io/src/commonMain/kotlin/space/kscience/dataforge/io/TaggedEnvelopeFormat.kt
@@ -3,8 +3,6 @@ package space.kscience.dataforge.io
 import io.ktor.utils.io.core.*
 import space.kscience.dataforge.context.Context
 import space.kscience.dataforge.context.Global
-import space.kscience.dataforge.io.IOFormat.Companion.META_KEY
-import space.kscience.dataforge.io.IOFormat.Companion.NAME_KEY
 import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.meta.enum
 import space.kscience.dataforge.meta.get
@@ -49,7 +47,7 @@ public class TaggedEnvelopeFormat(
         formatMeta: Meta,
     ) {
         val metaFormat = metaFormatFactory.build(this@TaggedEnvelopeFormat.io.context, formatMeta)
-        val metaBytes = metaFormat.toBinary(envelope.meta)
+        val metaBytes = Binary(envelope.meta,metaFormat)
         val actualSize: ULong = (envelope.data?.size ?: 0).toULong()
         val tag = Tag(metaFormatFactory.key, metaBytes.size.toUInt() + 2u, actualSize)
         output.writeBinary(tag.toBinary())
@@ -74,7 +72,7 @@ public class TaggedEnvelopeFormat(
 
         val metaBinary = input.readBinary(tag.metaSize.toInt())
 
-        val meta: Meta = metaFormat.readObject(metaBinary)
+        val meta: Meta = metaFormat.readObjectFrom(metaBinary)
 
         val data = input.readBinary(tag.dataSize.toInt())
 
@@ -89,7 +87,7 @@ public class TaggedEnvelopeFormat(
 
         val metaBinary = input.readBinary(tag.metaSize.toInt())
 
-        val meta: Meta = metaFormat.readObject(metaBinary)
+        val meta: Meta = metaFormat.readObjectFrom(metaBinary)
 
 
         return PartialEnvelope(meta, (version.tagSize + tag.metaSize).toInt(), tag.dataSize)
@@ -104,13 +102,6 @@ public class TaggedEnvelopeFormat(
     public enum class VERSION(public val tagSize: UInt) {
         DF02(20u),
         DF03(24u)
-    }
-
-    override fun toMeta(): Meta = Meta {
-        NAME_KEY put name.toString()
-        META_KEY put {
-            "version" put version
-        }
     }
 
     public companion object : EnvelopeFormatFactory {