+++ dataforge-workspace/src/commonMain/kotlin/space/kscience/dataforge/workspace/envelopeData.kt
@@ -4,19 +4,18 @@ import space.kscience.dataforge.data.Data
 import space.kscience.dataforge.data.await
 import space.kscience.dataforge.io.*
 import space.kscience.dataforge.misc.DFInternal
+import kotlin.reflect.typeOf
 
 /**
  * Convert an [Envelope] to a data via given format. The actual parsing is done lazily.
  */
 @OptIn(DFInternal::class)
-public fun <T : Any> Envelope.toData(format: IOFormat<T>): Data<T> {
-    return Data(format.type, meta) {
-        data?.readWith(format) ?: error("Can't convert envelope without data to Data")
-    }
+public inline fun <reified T : Any> Envelope.toData(format: IOReader<T>): Data<T> = Data(typeOf<T>(), meta) {
+    data?.readWith(format) ?: error("Can't convert envelope without data to Data")
 }
 
-public suspend fun <T : Any> Data<T>.toEnvelope(format: IOFormat<T>): Envelope {
+public suspend fun <T : Any> Data<T>.toEnvelope(format: IOWriter<T>): Envelope {
     val obj = await()
-    val binary = format.toBinary(obj)
+    val binary = Binary(obj, format)
     return SimpleEnvelope(meta, binary)
 }